Make calendar menu pop up when current date is clicked
